Create: Create Stock
Essentially, it starts with the CreateStockRequestDto.cs, where you format the data fields that are relevant to CREATE.

The StockMapper.cs's ToStockFromCreateDto() function converts the DTO (from user input) into an Entity.

The StockController.cs calls stockDto.ToStockFromCreateDto() to convert the DTO (from the user) to an entity, then saves it to the database.